requirements.txt
flask>=2.0
numpy
umap-learn
